+++ src/main/java/live/ioteatime/frontservice/dto/MonthlyElectricitiesDto.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class MonthlyElectricitiesDto {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private LocalDateTime time;
-    private long kwh;
+    private double kwh;
     private long bill;
     private List<DailyElectricityDto> dailyElectricityDtos;
 }
